requirements.txt
Pillow
numpy
tensorflow
scipy
kaggle